requirements.txt
aiohttp
aio-pika
beautifulsoup4
lxml
python-dotenv
